moving test over to new tfe test variable resource
